Functions

This page contains documentation for all exported functions.

Training

Call the main function bosip! to run the BOSIP procedure, which sequentially queries the expensive blackbox simulator to learn the parameter posterior efficiently.

BOSIP.bosip! — Function

bosip!(::BosipProblem; kwargs...)

Run the BOSIP method on the given BosipProblem.

The bosip! function is a wrapper for BOSS.bo!, which implements the underlying Bayesian optimization procedure.

Arguments

problem::BosipProblem: Defines the inference problem, together with all model hyperparameters.

Keywords

model_fitter::BOSS.ModelFitter: Defines the algorithm used to estimate the model hyperparameters.

acq_maximizer::BOSS.AcquisitionMaximizer: Defines the algorithm used to maximize the acquisition function in order to select the next evaluation point in each iteration.

term_cond::Union{<:BOSS.TermCond, <:BosipTermCond}: Defines the termination condition of the whole procedure.

options::BosipOptions: Can be used to specify additional miscellaneous options.

References

BOSS.bo!, BosipProblem, BosipAcquisition, BOSS.ModelFitter, BOSS.AcquisitionMaximizer, BOSS.TermCond, BosipTermCond, BosipOptions

Examples

See 'https://soldasim.github.io/BOSIP.jl/stable/example_lfi' for example usage.

Call the function estimate_parameters! to fit the model hyperparameters according to the current dataset. (One can also call bosip! with term_cond = IterLimit(0) to fit the hyperparameters without running any simulations. This will additionally only refit the model if the dataset changed since the last parameter estimation. In contrast, calling estimate_parameters! will always re-run the parameter estimation.)

BOSS.estimate_parameters! — Function

estimate_parameters!(::BossProblem, ::ModelFitter)

Estimate the model parameters & hyperparameters using the given model_fitter algorithm.

Keywords

options::BossOptions: Defines miscellaneous settings.

estimate_parameters!(::BosipProblem, ::ModelFitter)

Estimate the hyperparameters of the model. Uses the provided ModelFitter to fit the hyperparameters of the model according to the data stored in the BosipProblem.

Keywords

options::BosipOptions: Defines miscellaneous settings.

Analytically compute the optimal estimate of the distribution parameters according to the given data xs.

This function is a part of the optional API of the ProposalDistribution and may not be implemented for every distribution.

Call the function maximize_acquisition to obtain a promising candidate for the next simulation.

BOSS.maximize_acquisition — Function

x = maximize_acquisition(::BossProblem, ::AcquisitionMaximizer)

Maximize the given acquisition function via the given acq_maximizer algorithm to find the optimal next evaluation point(s).

Keywords

options::BossOptions: Defines miscellaneous settings.

x = maximize_acquisition(::BosipProblem, ::AcquisitionMaximizer)

Select parameters for the next simulation. Uses the provided AcquisitionMaximizer to maximize the acquisition function and find the optimal candidate parameters.

Keywords

options::BosipOptions: Defines miscellaneous settings.

Call the function eval_objective! to start a simulation run.

BOSS.eval_objective! — Function

eval_objective!(::BossProblem, x::AbstractVector{<:Real})

Evaluate the objective function and update the data.

Keywords

options::BossOptions: Defines miscellaneous settings.

eval_objective!(::BosipProblem, x::AbstractVector{<:Real})

Evaluate the blackbox simulation for the given parameters x.

Keywords

options::BosipOptions: Defines miscellaneous settings.

Parameter Posterior & Likelihood

This section contains function used to obtain the trained parameter posterior/likelihood approximations.

The approx_posterior function can be used to obtain the (un)normalized approximate posterior $p(x|z_o) \propto p(z_o|x) p(x)$ obtained by substituting the predictive means of the GPs directly as the discrepancies from the true observation.

BOSIP.approx_posterior — Function

approx_posterior(::BosipProblem; kwargs...)

Return the MAP estimation of the unnormalized approximate posterior $\hat{p}(z_o|x) p(x)$ as a function of $x$.

If normalize=true, the resulting posterior is approximately normalized.

The posterior is approximated by directly substituting the predictive means of the GPs as the discrepancies from the true observation and ignoring both the uncertainty of the GPs due to a lack of data and due to the simulator evaluation noise.

By using approx_posterior or posterior_mean one controls, whether to integrate over the uncertainty in the discrepancy estimate. In addition to that, by providing a ModelFitter{MAP} or a ModelFitter{BI} to bosip! one controls, whether to integrate over the uncertainty in the GP hyperparameters.

Keywords

normalize::Bool: If normalize is set to true, the evidence $\hat{p}(z_o)$is estimated by sampling and the normalized approximate posterior\hat{p}(z_o|x) p(x) / \hat{p}(z_o)` is returned instead of the unnormalized one.

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the parameter prior $p(x)$ as a column-wise matrix. Only has an effect if normalize == true.

samples::Int: Controls the number of samples used to estimate the evidence. Only has an effect if normalize == true and isnothing(xs).

See Also

posterior_mean, posterior_variance, approx_likelihood

BOSIP.log_approx_posterior — Function

log_approx_posterior(::BosipProblem)

Return the log of the unnormalized approximate posterior $\hat{p}(z_o|x) p(x)$ as a function of x.

See Also

approx_posterior, log_approx_likelihood, log_posterior_mean, log_posterior_variance,

The posterior_mean function can be used to obtain the expected value of the (un)normalized posterior $\mathbb{E}\left[p(x|z_o)\right] \propto \mathbb{E}\left[p(z_o|x)p(x)\right]$ obtained by analytically integrating over the uncertainty of the GPs and the simulator.

BOSIP.posterior_mean — Function

posterior_mean(::BosipProblem; kwargs...)

Return the expectation of the unnormalized posterior $\mathbb{E}[\hat{p}(z_o|x) p(x)]$ as a function of $x$.

If normalize=true, the resulting expected posterior is approximately normalized.

The returned function maps parameters x to the expected posterior probability density value integrated over the uncertainty of the GPs due to a lack of data and due to the simulator evaluation noise.

By using approx_posterior or posterior_mean one controls, whether to integrate over the uncertainty in the discrepancy estimate. In addition to that, by providing a ModelFitter{MAP} or a ModelFitter{BI} to bosip! one controls, whether to integrate over the uncertainty in the GP hyperparameters.

Keywords

normalize::Bool: If normalize is set to true, the evidence $\hat{p}(z_o)$is estimated by sampling and the normalized expected posterior\mathbb{E}[\hat{p}(z_o|x) p(x)]` is returned instead of the unnormalized one.

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the parameter prior $p(x)$ as a column-wise matrix. Only has an effect if normalize == true.

samples::Int: Controls the number of samples used to estimate the evidence. Only has an effect if normalize == true and isnothing(xs).

See Also

approx_posterior, posterior_variance, likelihood_mean

BOSIP.log_posterior_mean — Function

log_posterior_mean(::BosipProblem)

Return the log of the expectation of the unnormalized posterior $\mathbb{E}[\hat{p}(z_o|x) p(x)]$ as a function of $x$.

See Also

posterior_mean, log_likelihood_mean, log_approx_posterior, log_posterior_variance

The posterior_variance function can be used to obtain the variance of the (un)normalized posterior $\mathbb{V}\left[p(x|z_o)\right] \propto \mathbb{V}\left[p(z_o|x)p(x)\right]$ obtained by analytically integrating over the uncertainty of the GPs and the simulator.

BOSIP.posterior_variance — Function

posterior_variance(::BosipProblem; kwargs...)

Return the variance of the unnormalized posterior $\mathbb{V}[\hat{p}(z_o|x) p(x)]$ as a function of $x$.

If normalize=true, the resulting posterior variance is approximately normalized.

The returned function maps parameters x to the variance of the posterior probability density value estimate caused by the uncertainty of the GPs due to a lack of data and due to the simulator evaluation noise.

By providing a ModelFitter{MAP} or a ModelFitter{BI} to bosip! one controls, whether to compute the variance over the uncertainty in the GP hyperparameters as well.

Keywords

normalize::Bool: If normalize is set to true, the evidence $\hat{p}(z_o)$is estimated by sampling and the normalized posterior variance\mathbb{V}[\hat{p}(z_o|x) p(x) / \hat{p}(z_o)]` is returned instead of the unnormalized one.

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the parameter prior $p(x)$ as a column-wise matrix. Only has an effect if normalize == true.

samples::Int: Controls the number of samples used to estimate the evidence. Only has an effect if normalize == true and isnothing(xs).

See Also

approx_posterior, posterior_mean, likelihood_variance

BOSIP.log_posterior_variance — Function

log_posterior_variance(::BosipProblem)

Return the log of the variance of the unnormalized posterior $\mathbb{V}[\hat{p}(z_o|x) p(x)]$ as a function of $x$.

See Also

posterior_variance, log_likelihood_variance, log_posterior_mean, log_approx_posterior

The approx_likelihood function can be used to obtain the approximate likelihood $p(z_o|x)$ obtained by substituting the predictive means of the GPs directly as the discrepancies from the true observation.

BOSIP.approx_likelihood — Function

approx_likelihood(::BosipProblem)

Return the MAP estimation of the likelihood $\hat{p}(z_o|x)$ as a function of $x$.

The likelihood is approximated by directly substituting the predictive means of the GPs as the discrepancies from the true observation and ignoring both the uncertainty of the GPs due to a lack of data and due to the simulator evaluation noise.

By using approx_likelihood or likelihood_mean one controls, whether to integrate over the uncertainty in the discrepancy estimate. In addition to that, by providing a ModelFitter{MAP} or a ModelFitter{BI} to bosip! one controls, whether to integrate over the uncertainty in the GP hyperparameters.

See Also

likelihood_mean, likelihood_variance, approx_posterior

BOSIP.log_approx_likelihood — Function

log_approx_likelihood(::Likelihood, ::ModelPosterior)

Returns a function log_approx_like mapping $x$ to $log \hat{p}(z_o|x)$, with the following two methods:

log_approx_like(x::AbstractVector{<:Real}) -> ::Real

log_approx_like(X::AbstractMatrix{<:Real}) -> ::AbstractVector{<:Real}

The likelihood_mean function can be used to obtain the expected value of the likelihood $\mathbb{E}\left[p(z_o|x)\right]$ obtained by analytically integrating over the uncertainty of the GPs and the simulator.

BOSIP.likelihood_mean — Function

likelihood_mean(::BosipProblem)

Return the expectation of the likelihood approximation $\mathbb{E}[\hat{p}(z_o|x)]$ as a function of $x$.

The returned function maps parameters x to the expected likelihood probability density value integrated over the uncertainty of the GPs due to a lack of data and due to the simulator evaluation noise.

By using approx_likelihood or likelihood_mean one controls, whether to integrate over the uncertainty in the discrepancy estimate. In addition to that, by providing a ModelFitter{MAP} or a ModelFitter{BI} to bosip! one controls, whether to integrate over the uncertainty in the GP hyperparameters.

See Also

approx_likelihood, likelihood_variance, posterior_mean

BOSIP.log_likelihood_mean — Function

log_likelihood_mean(::Likelihood, ::ModelPosterior)

Returns a function log_like_mean mapping $x$ to $log \mathbb{E}[ \hat{p}(z_o|x) | GP ]$, with the following two methods:

log_like_mean(x::AbstractVector{<:Real}) -> ::Real

log_like_mean(X::AbstractMatrix{<:Real}) -> ::AbstractVector{<:Real}

The likelihood_variance function can be used to obtain the variance of the likelihood $\mathbb{V}\left[p(z_o|x)\right]$ obtained by analytically integrating over the uncertainty of the GPs and the simulator.

BOSIP.likelihood_variance — Function

likelihood_variance(::BosipProblem)

Return the variance of the likelihood approximation $\mathbb{V}[\hat{p}(z_o|x)]$ as a function of $x$.

The returned function maps parameters x to the variance of the likelihood probability density value estimate caused by the uncertainty of the GPs due to a lack of data and the uncertainty of the simulator due to the evaluation noise.

By providing a ModelFitter{MAP} or a ModelFitter{BI} to bosip! one controls, whether to compute the variance over the uncertainty in the GP hyperparameters as well.

See Also

approx_likelihood, likelihood_mean, posterior_variance

BOSIP.log_likelihood_variance — Function

log_likelihood_variance(::Likelihood, ::ModelPosterior)

Return a function log_like_var mapping $x$ to $log \mathbb{V}[ \hat{p}(z_o|x) | GP ]$, with the following two methods:

log_like_var(x::AbstractVector{<:Real}) -> ::Real

log_like_var(X::AbstractMatrix{<:Real}) -> ::AbstractVector{<:Real}

The approx_marginal_likelihood and marginal_likelihood_mean functions (and their log_ variants) return the per-dimension likelihood contributions $p(z_o^{(i)}|x)$, whose product recovers the corresponding full-likelihood quantity. They are only available for Marginalizable likelihoods.

BOSIP.approx_marginal_likelihood — Function

approx_marginal_likelihood(bosip::BosipProblem) -> ::Function

Return the MAP estimation of the per-dimension likelihoods $\hat{p}(z_o^{(i)}|x)$ as a function of $x$. The product of the returned values recovers approx_likelihood.

See approx_likelihood for details on the approximation approach and the difference between using approx_marginal_likelihood vs. marginal_likelihood_mean.

See Also

marginal_likelihood_mean, approx_likelihood, log_approx_marginal_likelihood

BOSIP.log_approx_marginal_likelihood — Function

log_approx_marginal_likelihood(::Likelihood, ::ModelPosterior)

Returns a function log_approx_ml mapping $x$ to the per-dimension $log \hat{p}(z_o^{(i)}|x)$, with the following two methods:

log_approx_ml(x::AbstractVector{<:Real}) -> ::AbstractVector{<:Real}

log_approx_ml(X::AbstractMatrix{<:Real}) -> ::AbstractMatrix{<:Real}

BOSIP.marginal_likelihood_mean — Function

marginal_likelihood_mean(bosip::BosipProblem) -> ::Function

Return the expectation of the per-dimension likelihoods $\mathbb{E}[\hat{p}(z_o^{(i)}|x)]$ as a function of $x$. The product of the returned values recovers likelihood_mean.

See likelihood_mean for details on the approximation approach and the difference between using approx_marginal_likelihood vs. marginal_likelihood_mean.

See Also

approx_marginal_likelihood, likelihood_mean, log_marginal_likelihood_mean

BOSIP.log_marginal_likelihood_mean — Function

log_marginal_likelihood_mean(::Likelihood, ::ModelPosterior)

Returns a function log_ml_mean mapping $x$ to the per-dimension $log \mathbb{E}[ \hat{p}(z_o^{(i)}|x) | GP ]$, with the following two methods:

log_ml_mean(x::AbstractVector{<:Real}) -> ::AbstractVector{<:Real}

log_ml_mean(X::AbstractMatrix{<:Real}) -> ::AbstractMatrix{<:Real}

The evidence function can be used to approximate the evidence $p(z_o)$ of a given posterior function by sampling. It is advisable to use this estimate only in low parameter dimensions, as it will require many samples to achieve reasonable precision on high-dimensional domains.

The evidence is the normalization constant needed to obtain the normalized posterior. The evidence function is used to normalize the posterior if one calls approx_posterior, posterior_mean, or posterior_variance with normalize=true.

BOSIP.evidence — Function

evidence(post, proposal; kwargs...)
evidence(post, proposal, domain_area; kwargs...)

Return the estimated normalization constant (evidence) $\hat{p}(z_o)$ of the unnormalized pdf provided as post.

Non-log wrapper around log_evidence.

Arguments

post: A function ::AbstractVector{<:Real} -> ::Real representing any unnormalized pdf, typically the unnormalized posterior $\hat{p}(z_o|x) p(x)$.

proposal: A multivariate distribution used to draw parameter samples. Typically the prior $p(x)$.

domain_area: The area of the parameter domain. Only required for the self-normalized estimator. See log_evidence.

Keywords

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the proposal as a column-wise matrix.

samples::Int: Controls the number of samples used to estimate the evidence. Only has an effect if isnothing(xs).

See Also

log_evidence

BOSIP.log_evidence — Function

log_evidence(logpost, proposal; kwargs...)
log_evidence(logpost, proposal, domain_area; kwargs...)

Return the log of the estimated normalization constant (evidence) $\log \hat{p}(z_o)$ of the unnormalized pdf provided as logpost.

The first method uses a standard importance-sampling Monte Carlo estimator:

$\hat{p}(z_o) = \frac{1}{N} \sum_i \frac{f(x_i)}{q(x_i)}, \quad x_i \sim q$

The second method uses a self-normalized importance-sampling Monte Carlo estimator:

$\hat{p}(z_o) = S \cdot \frac{\sum_i w_i f(x_i)}{\sum_i w_i}, \quad w_i = \frac{1}{q(x_i)}, \quad x_i \sim q$

where $f(x)$ is the unnormalized pdf and $S$ is the domain area. The self-normalized variant can reduce variance when the proposal $q$ does not exactly match the prior $p(x)$.

Arguments

logpost: A function ::AbstractVector{<:Real} -> ::Real representing the log of any unnormalized pdf, typically the unnormalized posterior $\log \hat{p}(z_o|x) p(x)$.

proposal: A multivariate distribution $q$ used to draw parameter samples. Typically the prior $p(x)$, in which case the importance weights cancel.

domain_area: The area (volume) $S$ of the parameter domain. Only required for the self-normalized estimator.

Keywords

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the proposal as a column-wise matrix.

samples::Int: Controls the number of samples used to estimate the evidence. Only has an effect if isnothing(xs).

See Also

evidence

The like and loglike functions evaluate the likelihood $p(z_o|δ)$ (and its log) for a given model output δ. The like_marginal and loglike_marginal functions return the per-dimension contributions (only for Marginalizable likelihoods).

BOSIP.like — Function

like(::Likelihood, δ::AbstractVector{<:Real}, [x::AbstractVector{<:Real}]) -> ::Real
like(l::Likelihood, Δ::AbstractMatrix{<:Real}, [X::AbstractMatrix{<:Real}]) -> ::AbstractVector{<:Real}

Return the likelihood of the observation given the model output δ.

BOSIP.like_marginal — Function

like_marginal(::Likelihood, δ::AbstractVector{<:Real}, [x::AbstractVector{<:Real}]) -> ::AbstractVector{<:Real}
like_marginal(l::Likelihood, Δ::AbstractMatrix{<:Real}, [X::AbstractMatrix{<:Real}]) -> ::AbstractMatrix{<:Real}

Return the per-dimension likelihoods of the observation given the model output δ. Only defined for Marginalizable likelihoods.

BOSIP.loglike — Function

loglike(::Likelihood, δ::AbstractVector{<:Real}, [x::AbstractVector{<:Real}]) -> ::Real
loglike(::Likelihood, Δ::AbstractMatrix{<:Real}, [X::AbstractMatrix{<:Real}]) -> ::AbstractVector{<:Real}

Return the log-likelihood of the observation given the proxy variable δ. Rarely, some Likelihoods may require the input parameters x to compute the log-likelihood as well.

For Marginalizable likelihoods the vector methods fall back on summing loglike_marginal (with or without x). For JointOnly likelihoods the x-independent vector method is implemented directly by the subtype, and the x-dependent one falls back on it. The matrix methods broadcast the vector methods over the columns (subtypes may override them for efficiency).

BOSIP.loglike_marginal — Function

loglike_marginal(like::Likelihood, δ::AbstractVector{<:Real}, [x::AbstractVector{<:Real}]) -> ::AbstractVector{<:Real}
loglike_marginal(like::Likelihood, Δ::AbstractMatrix{<:Real}, [X::AbstractMatrix{<:Real}]) -> ::AbstractMatrix{<:Real}

Return the per-dimension log-likelihoods of the observation given the proxy variable δ. The result is a vector with one entry per observation dimension, such that sum(loglike_marginal(like, δ)) ≈ loglike(like, δ).

The batched method returns a matrix where each column corresponds to one sample in Δ. Rarely, some Likelihoods may require the input parameters x to compute the log-likelihood as well.

Only defined for Marginalizable likelihoods, which implement the δ-only method; the rest is provided by the fallbacks below.

Acquisition Function

The function construct_acquisition can be used to obtain the acquisition function.

BOSS.construct_acquisition — Function

construct_acquisition(::AcquisitionFunction, ::BossProblem, ::BossOptions) -> (x -> ::Real)

Construct the given AcquisitionFunction for the given BossProblem.

The returned function acq(x::AbstractVector{<:Real}) = val::Real is maximized to select the next evaluation function of blackbox function in each iteration.

This method must be implemented for all subtypes of AcquisitionFunction.

Sampling from the Posterior

The sample_approx_posterior, sample_expected_posterior, and sample_posterior functions can be used to obtain approximate samples from the trained parameter posterior.

BOSIP.sample_approx_posterior — Function

xs, ws = sample_approx_posterior(bosip::BosipProblem, sampler::DistributionSampler, count::Int; kwargs...)

Sample count samples from the approximate posterior of the BosipProblem using the specified sampler. Return a column-wise matrix of the drawn samples.

Keywords

options::BosipOptions: Miscellaneous preferences. Defaults to BosipOptions().

See Also

sample_expected_posterior, sample_posterior, resample

BOSIP.sample_expected_posterior — Function

xs, ws = sample_approx_posterior(bosip::BosipProblem, sampler::DistributionSampler, count::Int; kwargs...)

Sample count samples from the expected posterior (i.e. the posterior mean) of the BosipProblem using the specified sampler. Return a column-wise matrix of the drawn samples.

Keywords

options::BosipOptions: Miscellaneous preferences. Defaults to BosipOptions().

See Also

sample_approx_posterior, sample_posterior, resample

BOSIP.sample_posterior — Function

sample_posterior(::DistributionSampler, logpost::Function, domain::Domain, count::Int; kwargs...)
sample_posterior(::DistributionSampler, loglike::Function, prior::MultivariateDistribution, domain::Domain, count::Int; kwargs...)

Sample count samples from the given posterior log-density function.

Keywords

options::BosipOptions: Miscellaneous preferences. Defaults to BosipOptions().

BOSIP.sample_posterior_pure — Function

xs = sample_posterior_pure(sampler::DistributionSampler, logpost::Function, domain::Domain, count::Int;
    supersample_ratio = 20,
)

Sample count samples from the posterior distribution defined by the logpost logpdf. Assures that the returned samples are "pure" (unweighted).

In case of a WeightedSampler, supersample_ratio × count samples are drawn, and subsequently down-sampled to count samples according to their weights.

BOSIP.resample — Function

xs = resample(xs::AbstractMatrix{<:Real}, ws::AbstractVector{<:Real}, count::Int)

Resample count samples from the given data set xs weighted by the given weights ws with replacement to obtain a new un-weighted data set.

Some data points may repeat in the resampled data set. Increasing the sample size of the initial data set may help to reduce the number of repetitions.

The sampling is performed via the Turing.jl package. The Turing.jl package is a quite heavy dependency, so it is not loaded by default. To sample from the posterior, one has to first load Turing.jl as using Turing, which will also compile the sample_posterior function.

Confidence Sets

This section contains function used to extract approximate confidence sets from the posterior. It is advised to use these approximations only with low-dimensional parameter domains, as they will require many samples to reach reasonable precision in high-dimensional domains.

The find_cutoff function can be used to estimate some confidence set of a given posterior function.

BOSIP.find_cutoff — Function

c = find_cutoff(target_pdf, xs, q)
c = find_cutoff(target_pdf, xs, ws, q)

Estimate the cutoff value c such that the set {x | post(x) >= c} contains q of the total probability mass.

The value c is estimated based on the provided samples xs sampled according to the target_pdf.

Alternatively, one can provide samples xs sampled according to some proposal_pdf with corresponding importance weights ws = target_pdf.(eachcol(xs)) ./ proposal_pdf.(eachcol(xs)).

See Also

approx_cutoff_area set_iou

The approx_cutoff_area function can be used to estimate the ratio of the area of a confidence set given by sum cutoff constant (perhaps found by find_cutoff) and the whole domain.

BOSIP.approx_cutoff_area — Function

V = approx_cutoff_area(target_pdf, xs, c)
V = approx_cutoff_area(target_pdf, xs, ws, c)

Approximate the ratio of the area where target_pdf(x) >= c relative to the whole support of target_pdf.

The are is estimated based on the provided samples xs sampled uniformly from the whole support of target_pdf.

Alternatively, one can provide samples xs sampled according to some proposal_pdf with corresponding importance weights ws = 1 ./ proposal_pdf.(eachcol(xs)).

See Also

find_cutoff set_iou

The set_iou function can be used to estimate the intersection-over-union (IoU) value between two sets.

BOSIP.set_iou — Function

iou = set_iou(in_A, in_B, x_prior, xs)

Approximate the intersection-over-union of two sets A and B.

The parameters in_A, in_B are binary arrays declaring which samples from xs fall into the sets A and B. The column-wise matrix xs contains the parameter samples. The samples have to be drawn from the common prior x_prior.

See Also

find_cutoff approx_cutoff_area

Evaluation Metrics

The calculate_metric function evaluates a DistributionMetric between the learned posterior and a reference, to assess the quality of the inference.

BOSIP.calculate_metric — Function

score = calculate_metric(::SampleMetric, true_samples::AbstractMatrix{<:Real}, approx_samples::AbstractMatrix{<:Real}; kwargs...)
score = calculate_metric(::PDFMetric, true_logpost::Function, approx_logpost::Function; kwargs...)

Evaluate the accuracy of the approximate parameter posterior.

Keywords

options::BosipOptions: Miscellaneous preferences. Defaults to BosipOptions().

Plotting Posterior Marginals

The functions plot_marginals_int and plot_marginals_kde are provided to visualize the trained posterior. Both functions create a matrix of figures containing the approximate marginal posteriors of each pair of parameters and the individual marginals on the diagonal.

The function plot_marginals_int approximates the marginals by numerical integration whereas the function plot_marginals_kde approximates the marginals by kernel density estimation.

BOSIP.plot_marginals_int — Function

using CairoMakie
fig = plot_marginals_int(::BosipProblem; kwargs...)

Create a matrix of plots displaying the marginal posterior distribution of each pair of parameters with the individual marginals of each parameter on the diagonal.

Approximates the marginals by numerically integrating the marginal integrals over a generated latin hypercube grid of parameter samples. The plots are normalized according to the plotting grid.

It is necessary to adapt the number of samples (defined through the lhc_grid_size keyword) according to the dimensionality of the parameter space.

Also provides an option to plot "marginals" of different functions by using the func and normalize keywords.

Kwargs

func::Function: Defines the function which is plotted. The plotted function f is defined as f = func(::BosipProblem). Reasonable options for func include approx_posterior, posterior_mean, posterior_variance etc.

plot_settings::PlotSettings: Settings for the plotting.

lhc_grid_size::Int: The number of samples in the generate LHC grid. The higher the number, the more precise marginal plots. This settings affects the computational cost of the plotting significantly.

matrix_ops::Bool: Set to false to disable the use of matrix operations for plotting the marginals is they are not supported for the given func. Disabling matrix operations can significantly hinder performance.

info::Bool: Set to false to disable prints.

display::Bool: Set to false to not display the figure. It is still returned.

See Also

plot_marginals_kde, compute_marginals_int, PlotSettings

BOSIP.plot_marginals_kde — Function

using CairoMakie
fig = plot_marginals_kde(::BosipProblem; kwargs...)

Create a matrix of plots displaying the marginal posterior distribution of each pair of parameters with the individual marginals of each parameter on the diagonal.

Approximates the marginals by kernel density estimation over parameter samples drawn by MCMC methods from the Turing.jl package. The plots are normalized according to the plotting grid.

It is necessary to adapt the number of samples (defined through the turing_options keyword) according to the dimensionality of the parameter space.

One should experiment with different kernels, lengthscales, and sample_counts to obtain a good approximation of the marginals. See keyword arguments below.

Kwargs

logpost_func::Function: A function, which takes the BosipProblem and returns a function x -> log p(x|z_obs) computing the log-posterior density. Reasonable options include log_posterior_mean and log_approx_posterior.

sampler::DistributionSampler: The sampler used to draw samples from the posterior. Defaults to the AMISSampler.

sample_count::Int: The number of samples to draw from the posterior for the KDE. The higher the number, the more precise marginal plots. This settings affects the computational cost of the plotting significantly.

kernel::Kernel: The kernel used in the KDE.

lengthscale::Union{<:Real, <:AbstractVector{<:Real}}: The lengthscale for the kernel used in the KDE. Either provide a single length-scale used for all parameter dimensions as a real number, or provide individual length-scales for each parameter dimension as a vector of real numbers.

plot_settings::PlotSettings: Settings for the plotting.

info::Bool: Set to false to disable prints.

display::Bool: Set to false to not display the figure. It is still returned.

See Also

plot_marginals_int, compute_marginals_kde, PlotSettings

The visual settings of the marginal plots can be controlled by providing an instance of PlotSettings as a keyword argument to either of the aforementioned functions.

BOSIP.PlotSettings — Type

PlotSettings(; kwargs...)

Aggregates all plot settings for the plot_marginals_int and plot_marginals_kde functions.

Kwargs

grid_size::Int64: The size of the grid used for plotting. This hyperparameter greatly impacts the computational cost of the plotting.

resolution::Int: The resolution of the individual plots in the marginal matrix.

param_labels::Union{Nothing, Vector{String}}: Labels used for the parameters in the plots. Defaults to nothing, in which case the default labels "x1,x2,..." are used.

plot_data::Bool: Controls whether to plot the datapoints queried from the simulation (including the initial data).

plot_samples::Bool: Only relevant to plot_marginals_kde. Controls whether the samples from the posterior are also plotted.

plot_pair_marginals::Bool: Set to false to skip plotting the pairwise parameter marginals in the off-diagonal of the marginal matrix.

plot_diagonal::Bool: Set to diagonal=false to skip plotting the univariate parameter marginals on the diagonal of the marginal matrix.

full_matrix::Bool: Set to full_matrix=false to only plot the marginals below the diagonal and skipped the redundant mirrored plots above the diagonal.

upper_triangle::Bool Switch between plotting the lower triangle (upper_triangle=false) or the upper triangle (upper_triangle=true) of the marginal matrix. Only has an effect if full_matrix=false.

x_true::Union{Nothing, AbstractVector{<:Real}}: The true parameter values to be indicated in the plots.

title::Union{Nothing, String}: A plot title.

plot_style::Symbol: The style of the 2d plots. Can be :heatmap, :contour, or :contourf.

colormap::Symbol: The colormap used for the 2d plots. We recommend trying :matter, :viridis, or :plasma.

Alternatively, you can precompute the PlotData via compute_marginals_int or compute_marginals_kde and create the marginal plots later via plot_marginals.

BOSIP.compute_marginals_int — Function

data = compute_marginals_int(::BosipProblem; kwargs...)

Same as plot_marginals_int, but only computes and returns the PlotData used for later plotting via plot_marginals instead of creating the plots immediately.

Intended use is to precompute the PlotData and serialize them (e.g. via JLD2.jl) for later plotting.

See the documentation of plot_marginals_int for the keyword arguments.

See Also

plot_marginals_int, plot_marginals], PlotData, PlotSettings

BOSIP.compute_marginals_kde — Function

data = compute_marginals_kde(::BosipProblem; kwargs...)

Same as plot_marginals_kde, but only computes and returns the PlotData used for later plotting via plot_marginals instead of creating the plots immediately.

Intended use is to precompute the PlotData and serialize them (e.g. via JLD2.jl) for later plotting.

See the documentation of plot_marginals_kde for the keyword arguments.

See Also

plot_marginals_kde, plot_marginals], PlotData, PlotSettings

BOSIP.plot_marginals — Function

fig = plot_marginals(data::PlotData; kwargs...)

Create the marginal plots from previously computed PlotData via compute_marginals_int or compute_marginals_kde.

Kwargs

plot_settings::PlotSettings: Settings for the plotting.

display::Bool: Set to false to not display the figure. It is still returned.

See Also

compute_marginals_int, compute_marginals_kde, PlotData, PlotSettings

BOSIP.PlotData — Type

PlotData

Stores all precomputed data required for plotting the marginals.

Instances of PlotData can be obtained via the compute_marginals_int and compute_marginals_kde functions and the plots can be created later via the plot_marginals function.

Direct instantiation of PlotData by the user is not intended.

Utils

The function approx_by_gauss_mix together with the structure GaussMixOptions can be used to obtain a Gaussian mixture approximation the provided probability density function.

BOSIP.approx_by_gauss_mix — Function

Approximate the given posterior by a Gaussian mixture.

Find all modes via Optimization.jl, then approximate each mode with a mutlivariate Gaussian with mean in the mode and variance according to the second derivation of the true posterior in the mode.

BOSIP.GaussMixOptions — Type

GaussMixOptions(; kwargs...)

Contains all hyperparameters for the function approx_by_gauss_mix.

Kwargs

algorithm: Optimization algorithm used to find the modes.

multistart::Int: Number of optimization restarts.

parallel::Bool: Controls whether the individual optimization runs are performed in paralell.

static_schedule::Bool: If static_schedule=true then the :static schedule is used for parallelization. This is makes the parallel tasks sticky (non-migrating), but can decrease performance.

autodiff::SciMLBase.AbstractADType: Defines the autodiff library used for the optimization. (Only relevant if a gradient-based optimizer is set as algorithm.)

cluster_ϵs::Union{Nothing, Vector{Float64}}: The minimum distance between modes. Modes which are too close to a "more important" mode are discarded. Also defines the minimum distance of a mode from a domain boundary.

rel_min_weight::Float64: The minimum pdf value of a mode to be considered relative to the highest pdf value among all found modes.

kwargs...: Other kwargs are passed to the optimization algorithm.

The x_dim and y_dim functions return the input/output dimensions of a BosipProblem. The get_subset function constructs the likelihood of a subset of the observation dimensions (used when BosipProblem.y_sets is set).

BOSS.x_dim — Function

x_dim(::BossProblem) -> Int

Return the input dimension of the problem.

Return the dimension of the proposal distribution.

BOSS.y_dim — Function

y_dim(::BossProblem) -> Int

Return the output dimension of the problem.

BOSIP.get_subset — Function

get_subset(::Likelihood, y_set::AbstractVector{<:Bool}) -> ::Likelihood

Construct the likelihood of the observation dimensions specified by y_set.