Functions

This page contains documentation for all exported functions.

Training

Call the main function bolfi! to run the BOLFI procedure, which sequentially queries the expensive blackbox simulator to learn the parameter posterior efficiently.

BOLFI.bolfi! — Function

bolfi!(::BolfiProblem; kwargs...)

Run the BOLFI method on the given BolfiProblem.

The bolfi! function is a wrapper for BOSS.bo!, which implements the underlying Bayesian optimization procedure.

Arguments

problem::BolfiProblem: Defines the inference problem, together with all model hyperparameters.

Keywords

acquisition::BolfiAcquisition: Defines the acquisition function.

model_fitter::BOSS.ModelFitter: Defines the algorithm used to estimate the model hyperparameters.

acq_maximizer::BOSS.AcquisitionMaximizer: Defines the algorithm used to maximize the acquisition function in order to select the next evaluation point in each iteration.

term_cond::Union{<:BOSS.TermCond, <:BolfiTermCond}: Defines the termination condition of the whole procedure.

options::BolfiOptions: Can be used to specify additional miscellaneous options.

References

BOSS.bo!, BolfiProblem, BolfiAcquisition, BOSS.ModelFitter, BOSS.AcquisitionMaximizer, BOSS.TermCond, BolfiTermCond, BolfiOptions

Examples

See 'https://soldasim.github.io/BOLFI.jl/stable/example_lfi' for example usage.

Parameter Distributions

This section contains function used to obtain the trained parameter posterior/likelihood approximations.

The approx_posterior function can be used to obtain the (un)normalized approximate posterior $p(\theta|y_o) \propto p(y_o|\theta) p(\theta)$ obtained by substituting the predictive means of the GPs directly as the discrepancies.

BOLFI.approx_posterior — Function

approx_posterior(::BolfiProblem; kwargs...)

Return the approximate posterior $\hat{p}(x|y_o)$ as a function of $x$.

The posterior is approximated by directly substituting the predictive means of the GPs as the discrepancies (and ignoring the variance of the GPs).

The unnormalized approximate posterior $\hat{p}(y_o|x) p(x)$ is returned by default.

Keywords

normalize::Bool: If normalize is set to true, the evidence $\hat{p}(y_o)$is estimated by sampling and the normalized approximate posterior\hat{p}(y_o|x) p(x) / \hat{p}(y_o)` is returned instead of the unnormalized one.

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the parameter prior $p(x)$ as a column-wise matrix. Only has an effect if normalize == true.

samples::Int: Controls the number of samples used to estimate the evidence. Only has an effect if normalize == true and isnothing(xs).

See Also

posterior_mean, posterior_variance, approx_likelihood

The posterior_mean function can be used to obtain the (un)normalized expected posterior $\mathbb{E}\left[p(\theta|y_o)\right] \propto \mathbb{E}\left[p(y_o|\theta)p(\theta)\right]$ obtained by analytically calculating the expectation of the posterior function over the uncertainty in the prediction of the discrepancies.

BOLFI.posterior_mean — Function

posterior_mean(::BolfiProblem; kwargs...)

Return the expectation of the posterior approximation $\mathbb{E}[\hat{p}(x|y_o)]$ as a function of $x$.

Keywords

normalize::Bool: If normalize is set to true, the evidence $\hat{p}(y_o)$is estimated by sampling and the normalized expected posterior\mathbb{E}[\hat{p}(y_o|x) p(x)]` is returned instead of the unnormalized one.

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the parameter prior $p(x)$ as a column-wise matrix. Only has an effect if normalize == true.

samples::Int: Controls the number of samples used to estimate the evidence. Only has an effect if normalize == true and isnothing(xs).

See Also

approx_posterior, posterior_variance, likelihood_mean

The posterior_variance function can be used to obtain the (un)normalized posterior variance $\mathbb{V}\left[p(\theta|y_o)\right] \propto \mathbb{V}\left[p(y_o|\theta)p(\theta)\right]$ caused by the uncertainty in the prediction of the discrepancies.

BOLFI.posterior_variance — Function

posterior_variance(::BolfiProblem; kwargs...)

Return the variance of the posterior approximation $\mathbb{V}[\hat{p}(x|y_o)]$ as a function of $x$.

The variance of the unnormalized posterior $\mathbb{V}[\hat{p}(y_o|x) p(x)]$ is returned by default.

Keywords

normalize::Bool: If normalize is set to true, the evidence $\hat{p}(y_o)$is estimated by sampling and the normalized posterior variance\mathbb{V}[\hat{p}(y_o|x) p(x) / \hat{p}(y_o)]` is returned instead of the unnormalized one.

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the parameter prior $p(x)$ as a column-wise matrix. Only has an effect if normalize == true.

samples::Int: Controls the number of samples used to estimate the evidence. Only has an effect if normalize == true and isnothing(xs).

See Also

approx_posterior, posterior_mean, likelihood_variance

The approx_likelihood function can be used to obtain the approximate likelihood $p(y_o|\theta)$ obtained by substituting the predictive means of the GPs directly as the discrepancies.

BOLFI.approx_likelihood — Function

approx_likelihood(::BolfiProblem; kwargs...)

Return the approximate likelihood $\hat{p}(y_o|x)$ as a function of $x$.

The likelihood is approximated by directly substituting the predictive means of the GPs as the discrepancies (and ignoring the variance of the GPs).

See Also

likelihood_mean, likelihood_variance, approx_posterior

The likelihood_mean function can be used to obtain the expected likelihood $\mathbb{E}\left[p(y_o|\theta)\right]$ obtained by analytically calculating the expectation of the likelihood function over the uncertainty in the prediction of the discrepancies.

BOLFI.likelihood_mean — Function

likelihood_mean(::BolfiProblem; kwargs...)

Return the expectation of the likelihood approximation $\mathbb{E}[\hat{p}(y_o|x)]$ as a function of $x$.

See Also

approx_likelihood, likelihood_variance, posterior_mean

The likelihood_variance function can be used to obtain the likelihood variance $\mathbb{V}\left[p(y_o|\theta)\right]$ caused by the uncertainty in the prediction of the discrepancies.

BOLFI.likelihood_variance — Function

likelihood_variance(::BolfiProblem; kwargs...)

Return the variance of the likelihood approximation $\mathbb{V}[\hat{p}(y_o|x)]$ as a function of $x$.

See Also

approx_likelihood, likelihood_mean, posterior_variance

The evidence function can be used to approximate the evidence $p(y_o)$ of a given posterior function by sampling. It is advisable to use this estimate only in low parameter dimensions, as it will require many samples to achieve reasonable precision on high-dimensional domains.

The evidence is the normalization constant needed to obtain the normalized posterior. The evidence function is used to normalize the posterior if one calls approx_posterior, posterior_mean, or posterior_variance with normalize=true.

BOLFI.evidence — Function

evidence(post, x_prior; kwargs...)

Return the estimated evidence $\hat{p}(y_o)$.

Arguments

post: A function ::AbstractVector{<:Real} -> ::Real representing the posterior $p(x|y_o)$.

x_prior: A multivariate distribution representing the prior $p(x)$.

Keywords

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the x_prior as a column-wise matrix.

samples::Int: Controls the number of samples used to estimate the evidence. Only has an effect if isnothing(xs).

Sampling

The sample_posterior function can be used to obtain approximate samples from the trained parameter posterior.

BOLFI.sample_posterior — Function

xs = sample_posterior(problem::BolfiProblem, options::TuringOptions)

Sample count samples from the learned posterior stored in problem.

The sampling is performed via the Turing.jl package. The Turing.jl package is a quite heavy dependency, so it is not loaded by default. To sample from the posterior, one has to first load Turing.jl as using Turing, which will also compile the sample_posterior function.

Confidence Sets

This section contains function used to extract approximate confidence sets from the posterior. It is advised to use these approximations only with low-dimensional parameter domains, as they will require many samples to reach reasonable precision in high-dimensional domains.

The find_cutoff function can be used to estimate some confidence set of a given posterior function.

BOLFI.find_cutoff — Function

post, c = find_cutoff(post, x_prior, q; kwargs...)

Estimate the cutoff value c such that the set {x | post(x) > c} contains q of the total probability mass.

The approximation is calculated by MC sampling from the x_prior.

The returned posterior post is unchanged.

Keywords

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the x_prior as a column-wise matrix.

samples::Int: Controls the number of samples used for the approximation. Only has an effect if isnothing(xs).

See Also

approx_cutoff_area set_iou

The approx_cutoff_area function can be used to estimate the ratio of the area of a confidence set given by sum cutoff constant (perhaps found by find_cutoff) and the whole domain.

BOLFI.approx_cutoff_area — Function

V = approx_cutoff_area(post, x_prior, c; kwargs...)

Approximate the ratio of the area where post(x) > c relative to the whole support of post(x).

The approximation is calculated by MC sampling from the x_prior.

The prior x_prior must support the whole support of post(x).

Keywords

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of samples from the x_prior as a column-wise matrix.

samples::Int: Controls the number of samples used for the approximation. Only has an effect if isnothing(xs).

See Also

find_cutoff set_iou

The set_iou function can be used to estimate the intersection-over-union (IoU) value between two sets.

BOLFI.set_iou — Function

iou = set_iou(in_A, in_B, x_prior, xs)

Approximate the intersection-over-union of two sets A and B.

The parameters in_A, in_B are binary arrays declaring which samples from xs fall into the sets A and B. The column-wise matrix xs contains the parameter samples. The samples have to be drawn from the common prior x_prior.

See Also

find_cutoff approx_cutoff_area