Data Types

Problem & Model

The BolfiProblem structure contains all information about the inference problem, as well as the model hyperparameters.

BOLFI.BolfiProblem — Type

BolfiProblem(X, Y; kwargs...)
BolfiProblem(::ExperimentData; kwargs...)

Defines the likelihood-free inference problem and stores all data.

Args

The initial data are provided either as two column-wise matrices X and Y with inputs and outputs of the simulator respectively, or as an instance of BOSS.ExperimentData.

Currently, at least one datapoint has to be provided (purely for implementation reasons).

Kwargs

f::Any: The simulation to be queried for data.

domain::Domain: The parameter domain of the problem.

acquisition::BolfiAcquisition: Defines the acquisition function.

model::SurrogateModel: The surrogate model to be used to model the proxy δ.

likelihood::Likelihood: The likelihood of the experiment observation z_o.

x_prior::MultivariateDistribution: The prior p(x) on the input parameters.

y_sets::Union{Nothing, Matrix{Bool}}: Optional parameter intended for advanced usage. The binary columns define subsets y_1, ..., y_m of the observation dimensions within y. The algorithm then trains multiple posteriors p(θ|y_1), ..., p(θ|y_m) simultaneously. The posteriors can be compared after the run is completed to see which observation subsets are most informative.

Likelihood

The abstract type Likelihood represents the likelihood distribution of the observation z_o.

BOLFI.Likelihood — Type

Represents the assumed likelihood of the experiment observation $z_o$.

Defining a Custom Likelihood

To define a custom likelihood, create a new subtype of Likelihood and implement the following API;

Each subtype of Likelihood should implement:

loglike(::Likelihood, z::AbstractVector{<:Real}) where z is the simulator output

log_approx_likelihood(::Likelihood, ::BolfiProblem, ::ModelPosterior)

log_likelihood_mean(::Likelihood, ::BolfiProblem, ::ModelPosterior)

Each subtype of Likelihood should implement at least one of:

log_sq_likelihood_mean(::Likelihood, ::BolfiProblem, ::ModelPosterior)

log_likelihood_variance(::Likelihood, ::BolfiProblem, ::ModelPosterior)

Additionally, the following method is also necessary to implement if BolfiProblem where !isnothing(problem.y_sets) is used:

get_subset(::Likelihood, y_set::AsbtractVector{<:Bool}):

The following additional methods are provided by default and need not be implemented:

like(::Likelihood, z::AbstractVector{<:Real})

The NormalLikelihood assumes that the observation z_o has been drawn from a Gaussian distribution with a known diagonal covariance matrix with the std_obs values on the diagonal. The simulator is used to learn the mean function.

BOLFI.NormalLikelihood — Type

NormalLikelihood(; z_obs, std_obs)

The observation is assumed to have been generated from a normal distribution as z_o \sim Normal(f(x), Diagonal(std_obs)). We can use the simulator to query z = f(x).

Kwargs

z_obs::Vector{Float64}: The observed values from the real experiment.

std_obs::Union{Vector{Float64}, Nothing}: The standard deviations of the Gaussian observation noise on each dimension of the "ground truth" observation. (If the observation is considered to be generated from the simulator and not some "real" experiment, provide std_obs = nothing` and the adaptively trained simulation noise deviation will be used in place of the experiment noise deviation as well. This may be the case for some toy problems or benchmarks.)

The LogNormalLikelihood assumes that the observation z_o has been drawn from a log-normal distribution with a known diagonal covariance matrix with the std_obs values on the diagonal. The simulator is used to learn the mean function.

BOLFI.LogNormalLikelihood — Type

LogNormalLikelihood(; z_obs, std_obs)

The observation is assumed to have been generated from a normal distribution as z_o \sim LogNormal(f(x), Diagonal(std_obs)). We can use the simulator to query z = f(x).

In many cases, one may want to take the logarithm of the output of the simulator. Meaning, if one has simulator z = sim(x), one would define f as y = f(x) = log(sim(x)). This way, the y values with high likelihood will have similar values to the z values.

Kwargs

z_obs::Vector{Float64}: The observed values from the real experiment.

std_obs::Vector{Float64}: The standard deviations of the LogNormal observation noise. (If the observation is considered to be generated from the simulator and not some "real" experiment, provide std_obs = nothing` and the adaptively trained simulation noise deviation will be used in place of the experiment noise deviation as well. This may be the case for some toy problems or benchmarks.)

The BinomialLikelihood assumes that the observation z_o has been drawn from a Binomial distribution with a known number trials. The simulator is used to learn the probability parameter p as a function of the input parameters. The expectation over this likelihood (in case one wants to use posterior_mean and/or posterior_variance) is calculated via simple numerical integration on a predefined grid.

BOLFI.BinomialLikelihood — Type

BinomialLikelihood(; z_obs, trials, kwargs...)

The observation is assumed to have been generated from a Binomial distribution as z_o \sim Binomial(trials, f(x)). We can use the simulator to query z = f(x).

The simulator should only return values between 0 and 1. The GP estimates are clamped to this range.

Kwargs

z_obs::Vector{Int64}: The observed values from the real experiment.

trials::Vector{Int64}: The number of trials for each observation dimension.

int_grid_size::Int64: The number of samples used to approximate the expected likelihood.

The ExpLikelihood assumes that the function f of the BolfiProblem already maps the parameters $x$ to the log-likelihood $\log p(z_o|y)$. Thus, the ExpLikelihood only exponentiates the surrogate model output $\delta$ to obtain the likelihood value.

BOLFI.ExpLikelihood — Type

ExpLikelihood(; kwargs...)

Assumes the model approximates the log-likelihood directly (as a scalar). Only exponentiates the model prediction.

Acquisition Function

The abstract type BolfiAcquisition represents the acquisition function.

BOLFI.BolfiAcquisition — Type

An abstract type for BOLFI acquisition functions.

Required API for subtypes of BolfiAcquisition:

Implement method (::CustomAcq)(::Type{<:UniFittedParams}, ::BolfiProblem, ::BolfiOptions) -> (x -> ::Real).

Optional API for subtypes of BolfiAcquisition:

Implement method (::CustomAcq)(::Type{<:MultiFittedParams}, ::BolfiProblem, ::BolfiOptions) -> (x -> ::Real). A default fallback is provided for MultiFittedParams, which averages individual acquisition functions for each sample.

The MaxVar can be used to solve LFI problems. It maximizes the posterior variance to select the next evaluation point.

BOLFI.MaxVar — Type

MaxVar()

Selects the new evaluation point by maximizing the variance of the posterior approximation.

BOLFI.LogMaxVar — Type

LogMaxVar()

Selects the new evaluation point by maximizing the log variance of the posterior approximation.

The LogMaxVar acquisition is functionally equivalent to MaxVar. Using MaxVar or LogMaxVar can be more/less suitable in different scenarios. Switching between the two can help with numerical stability.

The EIIG acquisition maximizes the Expected Integrated Information Gain. That is; it attempts to minimize the entropy of the current distribution over the possible parameter posteriors (which is implicitly given by the explicit surrogate model posterior).

BOLFI.EIIG — Type

EIIG(; kwargs...)

Selects new data point by maximizing the Expected Integrate Information Gain (EIIG).

The information gain is calculated as the mutual information between the new data point (a vector-valued random variable from a multivariate distribution given by the GPs) and the posterior approximation (a "random function" from a infinite-dimensional distribution).

Instead of calculating the information gain of the infinite-dimensional parameter posterior function values, EIIG computes the average information gain integrated over the domain.

The resulting mutual information between the new data point and the functional values of the approximate posterior is estimated by calculating the maximum mean discrepancy (MMD) between their joint and marginal distributions. This is also known as the Hilbert-Schmidt independence criterion (HSIC).

The random function (an infinite-dimensional random variable) representing the posterior is reduces to samples, which are integrated over. The samples are drawn from the x_proposal.

Kwargs

x_samples::Int64: The amount of samples used to approximate the integral over the parameter domain.

samples::Int64: The amount of samples drawn from the joint and marginal distributions to estimate the HSIC value.

x_proposal::MultivariateDistribution: This distribution is used to sample parameter samples used to numerically approximate the integral over the parameter domain.

y_kernel::Kernel: The kernel used for the samples of the new data point.

p_kernel::Kernel: The kernel used for the posterior function value samples.

The MWMV can be used to solve LFSS problems. It maximizes the "mass-weighted mean variance" of the posteriors given by the different sensor sets.

BOLFI.MWMV — Type

MWMV(; kwargs...)

The Mass-Weighted Mean Variance acquisition function.

Selects the next evaluation point by maximizing a weighted average of the variances of the individual posterior approximations given by different sensor sets. The weights are determined as the total probability mass of the current data w.r.t. each approximate posterior.

Keywords

samples::Int: The number of samples used to estimate the evidence.

Termination Condition

The abstract type BolfiTermCond represents the termination condition for the whole BOLFI procedure. Additionally, any BOSS.TermCond from the BOSS.jl package can be used with BOLFI.jl as well, and it will be automatically converted to a BolfiTermCond.

BOLFI.BolfiTermCond — Type

An abstract type for BOLFI termination conditions.

Implementing custom termination condition:

Create struct CustomTermCond <: BolfiTermCond

Implement method (::CustomTermCond)(::BolfiProblem) -> ::Bool

The most basic termination condition is the BOSS.IterLimit, which can be used to simply terminate the procedure after a predefined number of iterations.

BOLFI.jl provides two specialized termination conditions; the AEConfidence, and the UBLBConfidence. Both of them estimate the degree of convergence by comparing confidence regions given by two different approximations of the posterior.

BOLFI.AEConfidence — Type

AEConfidence(; kwargs...)

Calculates the q-confidence region of the expected and the approximate posteriors. Terminates after the IoU of the two confidence regions surpasses r.

Keywords

max_iters::Union{Nothing, <:Int}: The maximum number of iterations.

samples::Int: The number of samples used to approximate the confidence regions and their IoU ratio. Only has an effect if isnothing(xs).

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of parameter samples from the x_prior defined in BolfiProblem.

q::Float64: The confidence value of the confidence regions. Defaults to q = 0.95.

r::Float64: The algorithm terminates once the IoU ratio surpasses r. Defaults to r = 0.95.

BOLFI.UBLBConfidence — Type

UBLBConfidence(; kwargs...)

Calculates the q-confidence region of the UB and LB approximate posterior. Terminates after the IoU of the two confidence intervals surpasses r. The UB and LB confidence intervals are calculated using the GP mean +- n GP stds.

Keywords

max_iters::Union{Nothing, <:Int}: The maximum number of iterations.

samples::Int: The number of samples used to approximate the confidence regions and their IoU ratio. Only has an effect if isnothing(xs).

xs::Union{Nothing, <:AbstractMatrix{<:Real}}: Can be used to provide a pre-sampled set of parameter samples from the x_prior defined in BolfiProblem.

n::Float64: The number of predictive deviations added/substracted from the GP mean to get the two posterior approximations. Defaults to n = 1..

q::Float64: The confidence value of the confidence regions. Defaults to q = 0.8.

r::Float64: The algorithm terminates once the IoU ratio surpasses r. Defaults to r = 0.8.

Miscellaneous

The BolfiOptions structure can be used to define miscellaneous settings of BOLFI.jl.

BOLFI.BolfiOptions — Type

BolfiOptions(; kwargs...)

Stores miscellaneous settings.

Keywords

info::Bool: Setting info=false silences the algorithm.

debug::Bool: Set debug=true to print stactraces of caught optimization errors.

parallel_evals::Symbol: Possible values: :serial, :parallel, :distributed. Defaults to :parallel. Determines whether to run multiple objective function evaluations within one batch in serial, parallel, or distributed fashion. (Only has an effect if batching AM is used.)

callback::Union{<:BossCallback, <:BolfiCallback}: If provided, the callback will be called before the BO procedure starts and after every iteration.

The abstract type BolfiCallback can be derived to define a custom callback, which will be called once before the BOLFI procedure starts, and subsequently in every iteration.

For an example usage of this functionality, see the example in the package repository, where a custom callback is used to create the plots.

BOLFI.BolfiCallback — Type

If a callback cb of type BolfiCallback is defined in BolfiOptions, the method cb(::BolfiProblem; kwargs...) will be called in every iteration.

cb(problem::BolfiProblem;
    model_fitter::BOSS.ModelFitter,
    acq_maximizer::BOSS.AcquisitionMaximizer,
    term_cond::TermCond,                        # either `BOSS.TermCond` or a `BolfiTermCond` wrapped into `TermCondWrapper`
    options::BossOptions,
    first::Bool,
)

Samplers

The subtypes of DistributionSampler can be used to draw samples from the trained parameter posterior distribution.

BOLFI.DistributionSampler — Type

DistributionSampler

Subtypes of DistributionSampler are used to sample from a probability distribution.

Each subtype of DistributionSampler should implement:

sample_posterior(::DistributionSampler, posterior::Function, domain::Domain, count::Int; kwargs...) -> (X, ws)

Each subtype of DistributionSampler may additionally implement:

sample_posterior(::DistributionSampler, likelihood::Function, prior::MultivariateDistribution, domain::Domain, count::Int; kwargs...) -> (X, ws)

See also: PureSampler, WeightedSampler

BOLFI.PureSampler — Type

PureSampler <: DistributionSampler

A DistributionSampler which samples directly from the provided pdf, and always returns samples with uniform weights.

BOLFI.WeightedSampler — Type

WeightedSampler <: DistributionSampler

A DistributionSampler which does not sample directly from the pdf, but instead returns samples with non-uniform weights correcting for the sampling bias.

In particular, the following distribution samplers are currently provided.

BOLFI.RejectionSampler — Type

RejectionSampler(; kwargs...)

A sampler that uses trivial rejection sampling to draw samples from the posterior distribution.

Keywords

likelihood_maximizer::LikelihoodMaximizer: The optimizer used to find the maximum likelihood value.

BOLFI.TuringSampler — Type

TuringSampler <: DistributionSampler(; kwargs...)

Aggregates settings for the sample_posterior function, which uses the Turing.jl package.

Keywords

sampler::Any: The sampling algorithm used to draw the samples.

warmup::Int: The amount of initial unused 'warmup' samples in each chain.

chain_count::Int: The amount of independent chains sampled.

leap_size: Every leap_size-th sample is used from each chain. (To avoid correlated samples.)

parallel: If parallel=true then the chains are sampled in parallel.

Sampling Process

In each sampled chain;

The first warmup samples are discarded.

From the following leap_size * samples_in_chain samples each leap_size-th is kept.

Then the samples from all chains are concatenated and returned.

Total drawn samples: 'chaincount * (warmup + leapsize * samplesinchain)' Total returned samples: 'chaincount * samplesin_chain'

BOLFI.AMISSampler — Type

AMIS(; kwargs...)

Adaptive Metropolis Importance Sampling (AMIS) sampler for posterior distributions.

The sampler first aproximates the posterior distribution by a Laplace approximation centered on the maximum of the posterior, or with a Gaussian mixture model, and draws samples from it in the 0th iteration.

Afterwards, the AMIS algorithm is run for iters iterations with a simple Gaussian proposal distribution re-fitted in each iteration.

Keywords

iters::Int: Number of iterations of the AMIS algorithm.

proposal_fitter::DistributionFitter: The algorithm used to re-fit the proposal distribution in each iteration. Defaults to the AnalyticalFitter.

gauss_mix_options::Union{Nothing, GaussMixOptions}: Options for the Gaussian mixture approximation used for the 0th iteration. Defaults to nothing, which means the Laplace approximation is used instead.

Evaluation Metric

The subtypes of DistributionMetric can be used to evaluate the quality of the learned parameter posterior distribution.

BOLFI.DistributionMetric — Type

Subtypes of DistributionMetric are used to evaluate the quality of the posterior approximation.

The DistributionMetrics are grouped into two categories; SampleMetric and PDFMetric.

BOLFI.SampleMetric — Type

SampleMetric is a subtype of DistributionMetric that evaluates the quality of the posterior approximation based on samples drawn from the true and approximate posteriors.

Each subtype of SampleMetric should implement:

calculate_metric(::DistributionMetric, true_samples::AbstractMatrix{<:Real}, approx_samples::AbstractMatrix{<:Real}; kwargs...) -> ::Real

See also: DistributionMetric, PDFMetric

BOLFI.PDFMetric — Type

PDFMetric is a subtype of DistributionMetric that evaluates the quality of the posterior approximation based on the probability density functions (pdfs) of the true and approximate posteriors.

Each subtype of PDFMetric should implement:

calculate_metric(::DistributionMetric, true_post::Function, approx_post::Function; kwargs...) -> ::Real

See also: DistributionMetric, SampleMetric

In particular, the following metrics are currently provided.

BOLFI.MMDMetric — Type

MMDMetric(; kwargs...)

Measures the quality of the posterior approximation by sampling from the true posterior and the approximate posterior and calculating the Maximum Mean Discrepancy (MMD) between the two sample sets.

Keywords

kernel::Kernel: The kernel used to calculate the MMD. It is important to choose appropriate lengthscales for the kernel.

BOLFI.TVMetric — Type

TVMetric(; kwargs...)

Measures the quality of the posterior approximation by approximating the Total Variation (TV) distance based on a precomputed parameter grid.

Keywords

grid::Matrix{Float64}: The parameter grid used to approximate the TV integral.

ws::Vector{Float64}: The weights for the grid points. Should be 1 / q(x), where q(x) is the probability density function of the distribution used to sample the grid points. (1 / domain_area is appropriate for an evenly distributed grid)

References

[1] Gutmann, Michael U., and Jukka Cor. "Bayesian optimization for likelihood-free inference of simulator-based statistical models." Journal of Machine Learning Research 17.125 (2016): 1-47.

[2] Järvenpää, Marko, et al. "Efficient acquisition rules for model-based approximate Bayesian computation." (2019): 595-622.